Create sh_newdeathnotices.lua
Basically makes it so that weapon killicons are replaced with their print names. This looks prettier because most SWEPs don't have killicons.
Yeah yeah this is kinda taken from stronghold but it was pretty much 1:1 with the default gamemode file to begin with and I also adapted it to use net library.